requirements.txt
streamlit==1.37.1
numpy==1.23.5
Pillow==10.4.0
tensorflow==2.17.0
scipy==1.13.1
opencv-python==4.10.0
streamlit-webrtc==0.47.7
